+++ R/base-query.R
@@ -22,7 +22,7 @@ query_api <- function(url) {
 #' @noRd
 get_base_url <- function() {
 
-  base_url <- "https://eodhd.com/api/"
+  base_url <- "https://eodhd.com/api"
 
   return(base_url)